More batch rendering stuff and cleanup.

+++ src/main/java/Util/Time.java
@@ -2,7 +2,7 @@ package Util;
 
 public class Time {
 
-    public static float startTime = System.nanoTime();
+    public static long startTime = System.nanoTime();
 
     public static float getTime() {
         return (float) ((System.nanoTime() - startTime) * 1E-9);
